requirements.txt
bcrypt==3.1.7
cffi==1.14.0
alembic==1.4.2
click==7.1.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
passlib==1.7.2
pycparser==2.20
pyspark==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.
stripe==2.47.0
Flask_Bootstrap==3.3.7.1
WTForms==2.2.1
            
